star-burger/requirements.txt
django==3.2
django-debug-toolbar==3.2.1
Pillow==8.4.0
environs[django]==9.3.2
django-phonenumber-field==6.3.0
djangorestframework~=3.13.1
phonenumbers==8.12.55
requests==2.28.1
geopy==2.2.0
rollbar==0.16.*
psycopg2-binary==2.9.*
django-taggit==1.3.*
django-taggit-helpers==0.1.4
